sprint-4/step-1: студентики плакали, кололилсь, но начали типизировать
