+++ package.json
@@ -7,10 +7,6 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.37",
-    "@types/react": "^16.8.6",
-    "@types/react-dom": "^16.8.6",
     "@ya.praktikum/react-developer-burger-ui-components": "^1.13.1",
     "prop-types": "^15.8.1",
     "react": "^16.8.6",
@@ -22,7 +18,7 @@
     "react-scripts": "4.0.3",
     "redux": "^4.1.2",
     "redux-thunk": "^2.4.1",
-    "typescript": "^4.5.3",
+    "typescript": "^4.6.4",
     "uuid": "^8.3.2",
     "web-vitals": "^1.1.2"
   },
@@ -51,6 +47,11 @@
     ]
   },
   "devDependencies": {
+    "@types/jest": "^27.5.0",
+    "@types/node": "^17.0.31",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-router-dom": "^5.3.3",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
