+++ src/index.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 import './index.css'
-import App from './components/App/App.jsx'
+import App from './components/App/App.tsx'
 import reportWebVitals from './reportWebVitals'
 import { compose, createStore, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
