+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,28 +1,29 @@
-import React from 'react'
+import React, { ChangeEvent} from "react";
 import { Redirect, useHistory } from 'react-router-dom'
-import { useDispatch, useSelector } from 'react-redux'
+import { RootStateOrAny, useDispatch, useSelector } from "react-redux";
 import { Input, PasswordInput } from '@ya.praktikum/react-developer-burger-ui-components'
 import { CLEAR_LOGIN_STATE, loginUser} from '../../services/actions/user'
 import styles from './LoginPage.module.css'
 import AuthForm from '../../components/AuthForm/AuthForm'
 import { getCookie } from '../../utils/coockies'
+import { TUserState, HistoryWithFrom } from "../../utils/types";
 
 const LoginPage = () => {
-  const history = useHistory()
-  const userState = useSelector(store => store.user)
+  const history = useHistory<HistoryWithFrom>()
+  const userState: TUserState = useSelector((store: RootStateOrAny):TUserState => store.user)
   const [email, setEmail] = React.useState('')
   const [password, setPassword] = React.useState('')
   const dispatch = useDispatch()
 
-  const onEmailChange = (e) => {
+  const onEmailChange = (e: ChangeEvent<HTMLInputElement>) => {
     setEmail(e.target.value)
   }
 
-  const onPasswordChange = (e) => {
+  const onPasswordChange = (e: ChangeEvent<HTMLInputElement>) => {
     setPassword(e.target.value)
   }
 
-  const onSubmit = () => {
+  const onSubmit = (): void => {
     dispatch({type: CLEAR_LOGIN_STATE})
     dispatch(loginUser({email, password}))
     setEmail('')
